Fixing error parsing for ESS6, extracting target for exception.
