+++ graylog2-server/src/main/java/org/graylog2/indexer/IndexNotFoundException.java
@@ -31,5 +31,4 @@ public class IndexNotFoundException extends ElasticsearchException {
     public static IndexNotFoundException create(String errorMessage, String index) {
         return new IndexNotFoundException(errorMessage, Collections.singletonList("Index not found for query: " + index + ". Try recalculating your index ranges."));
     }
-
 }
